# apps/armi-runtime/src/armi_runtime/interfaces/creator_web_resources/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatorActivityPageResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CreatorActivityItemResponse"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 2048,
                "pattern": "v1\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "projection_version": {
            "const": "creator-activity.v2",
            "title": "Projection Version",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "projection_version",
          "items",
          "next_cursor"
        ],
        "title": "CreatorActivityPageResponse",
        "type": "object"
      },
...
      "CreatorMemoryPageResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CreatorMemoryItemResponse"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 2048,
                "pattern": "v1\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "projection_version": {
            "const": "creator-memory.v2",
            "title": "Projection Version",
            "type": "string"
          },
          "retrieval_kind": {
            "const": "creator_view",
            "title": "Retrieval Kind",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "projection_version",
          "retrieval_kind",
          "items",
          "next_cursor"
        ],
        "title": "CreatorMemoryPageResponse",
        "type": "object"
      },
...
      "ErrorDescriptorResponse": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ErrorCategoryValue"
          },
          "code": {
            "pattern": "[A-Z][A-Z0-9_]{2,127}",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error_instance_id": {
            "anyOf": [
              {
                "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Instance Id"
          }
        },
        "required": [
          "category",
          "code"
        ],
        "title": "ErrorDescriptorResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "LifeRecordKindValue": {
        "enum": [
          "activity",
          "conversation",
          "material",
          "memory",
          "relationship",
          "self_change"
        ],
        "type": "string"
      },
      "LifeRecordPageResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LifeRecordItemResponse"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 2048,
                "pattern": "v1\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "projection_version": {
            "const": "life-record-query.v2",
            "title": "Projection Version",
            "type": "string"
          },
          "retrieval_kind": {
            "enum": [
              "exact_query",
              "creator_view"
            ],
            "title": "Retrieval Kind",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "projection_version",
          "retrieval_kind",
          "items",
          "next_cursor"
        ],
        "title": "LifeRecordPageResponse",
        "type": "object"
      },
...
      "OtherHumanPartyRecordPageResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OtherHumanPartyRecordResponse"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 2048,
                "pattern": "v1\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "projection_version": {
            "const": "other-human-record.v1",
            "title": "Projection Version",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "projection_version",
          "items"
        ],
        "title": "OtherHumanPartyRecordPageResponse",
        "type": "object"
      },
...
      "RejectedOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptorResponse"
          },
          "message": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "occurred_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Occurred At",
            "type": "string"
          },
          "status": {
            "const": "rejected",
            "title": "Status",
            "type": "string"
          },
          "trace_id": {
            "pattern": "[0-9a-f]{32}",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "trace_id",
          "occurred_at",
          "message",
          "status",
          "error"
        ],
        "title": "RejectedOutcomeResponse",
        "type": "object"
      },
...
      "UnavailableOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptorResponse"
          },
          "message": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "occurred_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Occurred At",
            "type": "string"
          },
          "recovery_hint": {
            "anyOf": [
              {
                "maxLength": 4096,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Hint"
          },
          "status": {
            "const": "unavailable",
            "title": "Status",
            "type": "string"
          },
          "trace_id": {
            "pattern": "[0-9a-f]{32}",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "trace_id",
          "occurred_at",
          "message",
          "status",
          "error"
        ],
        "title": "UnavailableOutcomeResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "browserSessionBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ARMI Creator Interface",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/activities": {
      "get": {
        "operationId": "listCreatorActivities",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorActivityPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "List Creator Activities"
      }
    },
...
    "/v1/life-records": {
      "get": {
        "operationId": "queryCreatorLifeRecords",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LifeRecordKindValue"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "pattern": "v1\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifeRecordPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Query Creator Life Records"
      }
    },
...
    "/v1/memories": {
      "get": {
        "operationId": "listCreatorMemories",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "pattern": "v1\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorMemoryPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "List Creator Memories"
      }
    },
...
    "/v1/other-human-records": {
      "get": {
        "operationId": "listOtherHumanRecordParties",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "pattern": "v1\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtherHumanPartyRecordPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "List Other Human Record Parties"
      }
    },
...
  }
}